inclusión de lógica para categorias. Cambio de tipo de dato de ids en las clases de Interger a Long

+++ src/main/java/com/example/desafio_profesional_back/repositories/CategoriaRepository.java
@@ -4,9 +4,9 @@ import com.example.desafio_profesional_back.models.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-//marcar una clase como un componente de acceso a datos
+//Anotación para marcar una clase como un componente de acceso a datos
 @Repository
-public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
+public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
     // Encuentra una categoría por su título
     Categoria findByTituloIgnoreCase(String titulo);
 }
